chore: remove monorepo
