+++ package.json
@@ -1,7 +1,6 @@
 {
 	"name": "@pvtnbr/eslint-config",
 	"version": "0.0.0-semantic-release",
-	"private": true,
 	"license": "MIT",
 	"repository": "privatenumber/eslint-config",
 	"funding": "https://github.com/privatenumber/eslint-config?sponsor=1",
@@ -9,12 +8,9 @@
 		"name": "Hiroki Osame",
 		"email": "hiroki.osame@gmail.com"
 	},
-	"publishConfig": {
-		"access": "public"
-	},
 	"scripts": {
 		"lint": "eslint .",
-		"test": "pnpm test --parallel && :"
+		"test": "jest"
 	},
 	"husky": {
 		"hooks": {
@@ -27,20 +23,43 @@
 			"npm test"
 		]
 	},
+	"peerDependencies": {
+		"eslint": "^7.15.0"
+	},
+	"dependencies": {
+		"@typescript-eslint/eslint-plugin": "^4.30.0",
+		"@typescript-eslint/parser": "^4.30.0",
+		"confusing-browser-globals": "^1.0.10",
+		"eslint-import-resolver-typescript": "^2.4.0",
+		"eslint-plugin-eslint-comments": "^3.2.0",
+		"eslint-plugin-import": "^2.24.2",
+		"eslint-plugin-jsonc": "^1.6.0",
+		"eslint-plugin-no-use-extend-native": "^0.5.0",
+		"eslint-plugin-node": "^11.1.0",
+		"eslint-plugin-promise": "^5.1.0",
+		"eslint-plugin-react": "^7.25.1",
+		"eslint-plugin-react-hooks": "^4.2.0",
+		"eslint-plugin-regexp": "^1.0.0",
+		"eslint-plugin-unicorn": "^35.0.0",
+		"eslint-plugin-vue": "^7.7.0"
+	},
 	"devDependencies": {
 		"@types/eslint": "^7.28.0",
 		"@types/jest": "^27.0.1",
+		"@types/react": "^17.0.19",
 		"eslint": "^7.32.0",
 		"husky": "^4.3.8",
 		"jest": "^27.1.0",
 		"lint-staged": "^11.1.2",
+		"react": "^17.0.2",
 		"semantic-release": "^17.4.7",
-		"semantic-release-monorepo": "^7.0.5"
+		"semantic-release-monorepo": "^7.0.5",
+		"typescript": "^4.4.2"
 	},
 	"eslintConfig": {
-		"extends": "./packages/base",
+		"extends": ".",
 		"ignorePatterns": [
-			"packages/*/test/fixtures/*"
+			"test/fixtures/*"
 		]
 	}
 }
