--- packages/base/package.json
@@ -1,40 +1,0 @@
-{
-	"name": "@pvtnbr/eslint-config-base",
-	"version": "0.0.0-semantic-release",
-	"description": "Base ESLint configuration",
-	"license": "MIT",
-	"repository": {
-		"type": "git",
-		"url": "https://github.com/privatenumber/eslint-config.git",
-		"directory": "packages/base"
-	},
-	"funding": "https://github.com/privatenumber/eslint-config?sponsor=1",
-	"author": {
-		"name": "Hiroki Osame",
-		"email": "hiroki.osame@gmail.com"
-	},
-	"files": [
-		"index.js",
-		"rules"
-	],
-	"publishConfig": {
-		"access": "public"
-	},
-	"scripts": {
-		"test": "jest"
-	},
-	"peerDependencies": {
-		"eslint": "^7.15.0"
-	},
-	"dependencies": {
-		"confusing-browser-globals": "^1.0.10",
-		"eslint-plugin-eslint-comments": "^3.2.0",
-		"eslint-plugin-import": "^2.24.2",
-		"eslint-plugin-jsonc": "^1.6.0",
-		"eslint-plugin-no-use-extend-native": "^0.5.0",
-		"eslint-plugin-node": "^11.1.0",
-		"eslint-plugin-promise": "^5.1.0",
-		"eslint-plugin-regexp": "^1.0.0",
-		"eslint-plugin-unicorn": "^28.0.2"
-	}
-}
