--- packages/base/rules/regexp.js
@@ -1,5 +1,0 @@
-module.exports = {
-	extends: [
-		'plugin:regexp/recommended',
-	],
-};
