--- packages/react/package.json
@@ -1,37 +1,0 @@
-{
-	"name": "@pvtnbr/eslint-config-react",
-	"version": "0.0.0-semantic-release",
-	"description": "React ESLint configuration",
-	"license": "MIT",
-	"repository": {
-		"type": "git",
-		"url": "https://github.com/privatenumber/eslint-config.git",
-		"directory": "packages/react"
-	},
-	"funding": "https://github.com/privatenumber/eslint-config?sponsor=1",
-	"author": {
-		"name": "Hiroki Osame",
-		"email": "hiroki.osame@gmail.com"
-	},
-	"files": [
-		"index.js"
-	],
-	"publishConfig": {
-		"access": "public"
-	},
-	"scripts": {
-		"test": "jest"
-	},
-	"peerDependencies": {
-		"eslint": "^7.15.0"
-	},
-	"dependencies": {
-		"@pvtnbr/eslint-config-typescript": "*",
-		"eslint-plugin-react": "^7.23.2",
-		"eslint-plugin-react-hooks": "4.2.0"
-	},
-	"devDependencies": {
-		"@types/react": "^16.8.24",
-		"react": "^17.0.2"
-	}
-}
