--- packages/typescript/package.json
@@ -1,37 +1,0 @@
-{
-	"name": "@pvtnbr/eslint-config-typescript",
-	"version": "0.0.0-semantic-release",
-	"description": "TypeScript ESLint configuration",
-	"license": "MIT",
-	"repository": {
-		"type": "git",
-		"url": "https://github.com/privatenumber/eslint-config.git",
-		"directory": "packages/typescript"
-	},
-	"funding": "https://github.com/privatenumber/eslint-config?sponsor=1",
-	"author": {
-		"name": "Hiroki Osame",
-		"email": "hiroki.osame@gmail.com"
-	},
-	"files": [
-		"index.js"
-	],
-	"publishConfig": {
-		"access": "public"
-	},
-	"scripts": {
-		"test": "jest"
-	},
-	"peerDependencies": {
-		"eslint": "^7.15.0"
-	},
-	"dependencies": {
-		"@pvtnbr/eslint-config-base": "*",
-		"@typescript-eslint/eslint-plugin": "^4.30.0",
-		"@typescript-eslint/parser": "^4.30.0",
-		"eslint-import-resolver-typescript": "^2.4.0"
-	},
-	"devDependencies": {
-		"typescript": "^4.4.2"
-	}
-}
