--- packages/typescript/test/fixtures/tsconfig.json
@@ -1,3 +1,0 @@
-{
-	"include": ["."]
-}
