+++ rules/es6.js
