+++ rules/promise.js
