+++ rules/regexp.js
@@ -1,0 +1,3 @@
+module.exports = {
+	extends: 'plugin:regexp/recommended',
+};
