+++ rules/variables.js
