+++ test/fixtures/InputComponent.tsx
