+++ test/fixtures/pass.js
