+++ test/fixtures/pass.ts
