+++ test/fixtures/tsconfig.json
@@ -1,6 +1,7 @@
 {
 	"compilerOptions": {
-		"esModuleInterop": true
+		"esModuleInterop": true,
+		"jsx": "react"
 	},
 	"include": ["."]
 }
